Fix off-by-one error
When trying to calculate the RMSDs involving the last
conformation, an offset one past the end of the rmsd vector
was being used. This is probably OK as it is never written
to (there are zero RMSDs to calculate as it is a triangular
matrix) but nevertheless is both slightly inefficient and
triggers a glibc assertion failure.